fix: F14 collection UX hotfixes — internal nav, price display, infinite scroll
- Collection cards stay internal (no Scryfall redirect for unlinked cards)
- Added Scryfall + LigaMagic external links inside CardDetail page
- Fixed price display: Decimal→float schema fix + defensive formatBRL
- Replaced Load More button with IntersectionObserver infinite scroll
- 715 backend tests (91.80% cov), 212 frontend tests, all green

+++ src/api/schemas/cards.py
@@ -14,7 +14,7 @@ class CardSummary(BaseModel):
     name_pt: str | None = None
     set_code: str | None = None
     collector_number: str | None = None
-    latest_price: Decimal | None = None
+    latest_price: float | None = None
 
     model_config = ConfigDict(from_attributes=True)
 
@@ -35,7 +35,7 @@ class CardDetail(BaseModel):
     name_pt: str | None = None
     set_code: str | None = None
     collector_number: str | None = None
-    latest_price: Decimal | None = None
+    latest_price: float | None = None
     source_cards: list[SourceCardSchema] = []
     created_at: datetime
     updated_at: datetime
